src/main/java/concuroo/nodes/expression/literalExpression/StringLiteral.java
package concuroo.nodes.expression.literalExpression;

import concuroo.nodes.expression.LiteralExpression;

public class StringLiteral implements LiteralExpression {

    private String value;

    public StringLiteral(String value) {
        this.value = value;
    }

    @Override
    public Object getValue() {
        return value;
    }

    @Override
    public void setValue(Object value) {
        this.value = value.toString();
    }

    @Override
    public String getLiteral() {
        return this.value;
    }
}